Move functions in PyCapnp.hpp to separate cpp file
