+++ src/nupic/py_support/PyCapnp.hpp
@@ -99,25 +99,8 @@ namespace nupic
     return proto;
   }
 
-  PyObject* getPyReader(capnp::DynamicStruct::Reader reader)
-  {
-    if (!initialized) {
-      initCapnpToPycapnp();
-      initialized = true;
-    }
-    py::Ptr parent(Py_None);
-    return createReader(reader, parent);
-  }
-
-  PyObject* getPyBuilder(capnp::DynamicStruct::Builder builder)
-  {
-    if (!initialized) {
-      initCapnpToPycapnp();
-      initialized = true;
-    }
-    py::Ptr parent(Py_None);
-    return createBuilder(builder, parent);
-  }
+  PyObject* getPyReader(capnp::DynamicStruct::Reader reader);
+  PyObject* getPyBuilder(capnp::DynamicStruct::Builder builder);
 
 }  // namespace nupic
